+++ docs/about/release-notes.rst
@@ -17,6 +17,7 @@ Features
 * Added support for passing any dict-like objects or iterables of tuples as ``coords``, ``attrs``, and ``masks`` arguments to initializers of :class:`scipp.DataArray` and :class:`scipp.Dataset` `#2603 <https://github.com/scipp/scipp/pull/2603>`_.
 * Added support for passing any dict-like objects or iterables of tuples as the ``data`` argument to the initializer of :class:`scipp.Dataset` `#2603 <https://github.com/scipp/scipp/pull/2603>`_.
 * Added support for ``DataArray`` and ``Dataset`` in more reduction operations (e.g. :func:`scipp.max`) `#2600 <https://github.com/scipp/scipp/pull/2600>`_.
+* Added support for masks in reductions of binned data except in :func:`scipp.mean` `#2608 <https://github.com/scipp/scipp/pull/2608>`_.
 * Added more reduction operations to the ``.bins`` property (e.g. :func:`scipp.Bins.max`) `#2608 <https://github.com/scipp/scipp/pull/2608>`_.
 * Added support for ``DataArray`` and ``Dataset`` in reduction operations (e.g. :func:`scipp.max`) `#2600 <https://github.com/scipp/scipp/pull/2600>`_.
 * Reduce effect of rounding errors when converting units `#2607 <https://github.com/scipp/scipp/pull/2607>`_.